fix: move archignore tests to test directory for architecture compliance
- Move test files from src/ to test/ directory (proper test organization)
- Update import paths in tests
- Fixes architecture rule violations (tests shouldn't be in src/)
- All tests now pass
